Implement Phase 6 mating handshake: require sustained contact for reproduction
Previously, sexual reproduction triggered instantly when two organisms with
MateIntent > 0.5 were in proximity. This allowed brief or accidental contacts
to produce offspring, which doesn't match realistic mating behavior.

The new dwell-time handshake requires both organisms to maintain:
- MateIntent above threshold (0.5)
- Energy above minimum ratio (35%)
- Surface-to-surface contact (≤5 world units)
- Mutual partner lock (tracking each other's entity ID)

...for 60 consecutive ticks before reproduction occurs. If any condition fails,
progress resets to zero and the handshake must restart.

Changes:
- Rewrite BreedingSystem.Update with 4-phase approach: update existing
  handshakes, start new handshakes, complete mature handshakes, handle asexual
- Add MateProgress and MatePartnerID tracking (fields already existed but unused)
- Add canStartHandshake and computeSurfaceDistance helpers
- Remove obsolete isEligible and isCompatibleForSexual functions
- Change MatePartnerID type from uint64 to uint32 to match ECS entity ID
- Precompute OBB center and body radius in breeder struct for efficiency

+++ components/components.go
@@ -87,7 +87,7 @@ type Organism struct {
 	// Attack state
 	AttackCooldown int32  // Ticks until can attack again
 	MateProgress   int32  // Ticks in mating contact
-	MatePartnerID  uint64 // Entity ID of current mating partner
+	MatePartnerID  uint32 // Entity ID of current mating partner (0 = none)
 
 	// Damage awareness (set by feeding system, decays each tick)
 	BeingEaten float32 // 0-1, how intensely this organism is being eaten
